backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
supabase==2.9.1
google-generativeai==0.8.5
gunicorn==23.0.0
python-dotenv==1.2.1
